Fix up special casing for closing review requests.
In the case of the close description field, the way that the review request
editor would call `reviewRequest.close` instead of setting a field on the model
itself was pretty hacky. The implementation of `setDraftField` would look at
whether it was setting the `closeDescription` and then special-case its
behavior.

Now that we have JavaScript-side implementations for the built-in fields, we
don't need this anymore. I've moved the special casing into the
`CloseDescriptionFieldView`.

This refactor will also make things easier down the road for new field types.

Testing done:
- Tested setting the close description with both regular and Markdown text.
- Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/9150/