Use the search index for the search resource when available
When indexing is enabled, we can use it to perform better full text
search for review requests and users. Since review groups are not
indexed, we still have to query the database for them even if search is
enabled. This does have the side effect that review requests published
by a user will show up in a search for that user's username, but that is
a convenient benefit.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/9116/