Add new VERIFY_* issue states to comment models.
This change adds the new plumbing for two new states for comment issues, which
are an intermediate step between "Open" and either "Fixed" or "Dropped". In
this state, the initial author of the comment will need to acknowledge the
change before it's made permanent. This will only be used when the comment
author explicitly requests that the issue will require verification.

Testing done:
- Ran unit tests.
- Used in conjunction with future changes.

Reviewed at https://reviews.reviewboard.org/r/9085/