Add display for issue verification.
This change adds UI to display issues waiting for verification to the comment
issue bar, the issue summary table, and the dashboard. While this will
currently display issues as "Waiting for verification" if they're in that
state, it does not actually show or handle the action for verifying it. That
will come in a later change.

Testing done:
- Ran unit tests.
- Used in conjunction with future UI change to add behavior.

Reviewed at https://reviews.reviewboard.org/r/9169/