Add an icon for comment issue verification.
This icon is used to represent an issue that is pending verification by
the reviewer. It has the same shape as an open issue, with a more
subdued (but still present) color, and arrows pointing in opposite
directions to indicate a back-and-forth between reviewer and developer.

A `.rb-icon-issue-verifying` CSS class has been added for using this icon.

Testing Done:
Tested in the browser by replacing the icon's CSS class on an open issue.

Reviewed at https://reviews.reviewboard.org/r/9095/