Convert page-related unit tests to ES6 and clean up some existing tests.
This converts the JavaScript unit tests for `RB.DiffViewerPageView` and
`RB.ReviewablePageView` to ES6, in preparation for a new set of tests
coming in.

The tests for `RB.ReviewRequestPageView` have been fixed to use
`function() {}` instead of `() => {}` for Jasmine tests, since the
former is recommended (due to Jasmine populating `this` with state, and
the ability for test suites to store their own local state in Jasmine's
`this`). It also fixes a variable name (`page` to `pageView`), which
will help with future tests.

Testing Done:
All unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9162/