Split off the common comment templates into a new template file.
All comments on a review have the same basic structure, which is
repeated for each type of comment. We've kept it this way for a while in
an attempt to avoid the overhead of including another template for each
comment, but it's getting difficult to update those, and we're pretty
good at caching these days.

This change splits off the common bits into a new
`reviews/entries/_review_comment.html` template, which handles
everything but the context above the comment. Rather than going with a
more complex solution of subclassing a base comment template for the
context, or having Python-side comment renderers, we're just having the
main `_review_body.html` template define a variable with the context
before including the common template, which works fine and keeps it
maintainable and a little faster.

Testing Done:
Tested every type of comment. Made sure each displayed correctly and
could be replied to.

Reviewed at https://reviews.reviewboard.org/r/9183/