Fix a display regression for page entries in mobile mode.
The move to the new entry system and styling for the review request page
regressed the presentation of the entry boxes a bit. The callout arrows
remained in mobile mode and the top of the box was flush with the
avatar.

This change fixes these regressions by bringing back the rules for
turning off the callouts and adding back the margin.

Testing Done:
Checked that the reviews looked correct again in mobile mode, and that the
styles remained correct in desktop mode.

Reviewed at https://reviews.reviewboard.org/r/9186/