Clean up some regressed visuals in the diff viewer.
Some of the recent template and style work made bits of the diff viewer
look a bit ugly. The issue summary table was always present (but empty),
and the diff entries being loaded had a white background behind the
spinner.

This change makes the issue summary table's display conditional,
enabling it only for the review request page. It also changes the look
of the diff entries to have the spinner before the filename, creating a
nice visual consistency and making the entries look nice and slim.

It also moves some padding from `.sidebyside.loading` into its `tbody`,
to avoid an issue with the wrong color appearing along the edges.

Testing Done:
Viewed the diff viewer page and saw that the issue summary table was no
longer there, and that the diff placeholder entries looked a lot nicer.

Made sure that the issue summary table was visible on the main review
request page.

Reviewed at https://reviews.reviewboard.org/r/9187/