Fix doc generation for the extension resources
The docs for the extension resource were still pointing at the old
Djblets extension resource instead of the one we provide in RB. This was
causing building the docs to crash.

In addition, our extension resource was not getting the
`required_features` attribute due to that attribute not being on the
`RBResourceMixin` but being on the `WebAPIResource` base class. The
attribute has been moved to the mixin, fixing another crash.

Testing Done:
With this patch applied, building the docs no longer crashes.

Reviewed at https://reviews.reviewboard.org/r/9185/