Set uploaded media files to download when accessed in a browser.
When accessing the URL for an uploaded media file, browsers will attempt
to view that media file inline. This can be dangerous, as some files
(such as HTML and SVG) may contain JavaScript that would run on the
server's origin (if hosted by the same server running Review Board).
This would allow them to make use of any logged in user sessions,
invoking the API on the user's behalf.

To avoid this, we can instruct the browser to download the contents of
the file directly, rather than viewing inline. This requires a
particular header to be passed to the browser when downloading the file,
which needs configuration in the web server. We can supply this via the
`.htaccess` files we generate in rb-site, which will work fine if using
a standard Apache install.

Not all sites use Apache alone. Some use a CDN (which is not affected by
this problem), while others use a server like nginx. As we don't
currently provide nginx configuration, and there's no equivalent of a
`.htaccess` file, we currently don't do anything for those users. We'll
have to rely on users paying attention to security announcements and to
follow the instructions.

We do, however, have a new security check in the admin UI, which will
test to see if the files are being served up with the right header,
allowing users to test their configuration changes. Some cleanups were
made to the security checking code to better support file-related
checks.

To help ensure that our security checks pass when in development mode,
we now use a custom `serve()` function that sets the appropriate headers
for all uploaded media files.

Testing Done:
Unit tests pass.

Tested uploading a malicious SVG file locally and confirming that when I
tried to access it via the file's URL, the file downloaded instead of
rendering in the browser.

Tested the same thing with the new configuration on an Apache server. Saw
that once again, the file downloaded instead of rendering inline.

Tested that the files could still be viewed properly in the review UI,
without any issues at all.

Reviewed at https://reviews.reviewboard.org/r/9159/