Add siteconfig var and button for read-only mode.
Read-only mode is a new feature that allows users to view Review Board while
the site is being maintained or upgraded.

This change adds a new toggle for read-only mode in Admin -> General Settings.
This connects to a new attribute to siteconfig called `site_read_only`.
Toggling read-only mode will toggle the attribute to true/false.

Testing Done:
In admin panel set new toggle to on/off and check that site_read_only in
siteconfig is set to true or false depending on whether the toggle is checked
or not.

Reviewed at https://reviews.reviewboard.org/r/8648/