Respond to API requests with 503 when read-only mode.
Read-only mode is a setting an admin can enable to prevent writes to the
database. This can be used when the site is under maintenence or being
upgraded. The previous commit added the siteconfig attribute and a checkbox to
toggle the mode. This commit adds a new 503 error to Review Board's WebAPI.
Furthermore, all requests that hit RB's API using `PUT`, `POST`, `DELETE`
methods while read-only mode is on will return this new 503 error (assuming the
user is not a superuser).

Testing done:
- Added new Python unit tests that send API requests as superuser and
  non-superuser.
- Ran the Python unit test suite.
- Check that modifying/deleting review requests and comments works properly

Reviewed at https://reviews.reviewboard.org/r/8657/