Include and make use of update timestamps for all entries.
Entries for the review request page previously only had a single
timestamp, representing the time the entry was first added, and
dictating the order of the page. The new code for handling updates to
entries on review requests checked that timestamp, and while this
appeared to work in a developer server environment (possibly a timezone
issue, which is still being looked at), it wasn't working in production.

This change introduces a new "updated" timestamp, representing when an
update was made to the entry. The existing timestamp is now an "added"
timestamp, used for order and "new entry" calculations. The "updated"
timestamp factors in things like replies to reviews and state changes
for status updates. The update checks make use of this now to determine
whether the entry's new HTML and metadata should be loaded onto the
page.

Unit tests were also added for all the new timestamp logic, and to
ensure update processing is now handled appropriately when there are
updates to existing entries.

Testing Done:
Tested with pending -> resolved status updates. Checked the timestamps
of the updates coming in and made sure that the `addedTimestamp` didn't
change, but that `updatedTimestamp` did.

Manually checked the added/updated timestamps of other entries in
response to changes I made.

All unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/9195/