Fix a handful of failures and regressions in our docs.
The documentation no longer built, due to a series of missing files and
fixtures, incorrect metadata, and bad references.

This change fixes all that by:

* Adding fixtures for issue verification state, OAuth2 apps/tokens, and
  user-based file attachments
* Adding missing API error and resource doc files
* Adding missing terms to the glossary
* Fixing metadata issues in new API resources
* Fixing the fixture dump script to include the new apps
* Fixing syntax errors in docstrings
* Fixing anchor names

It also fixes some bad filenames generated for the OAuth2 resources. The
uncamelcasing tried using "o-auth" instead of "oauth" in the filenames.
To fix this, support has been added for allowing names to be remapped as
needed in the config file.

Testing Done:
Built the docs. There were still a couple of errors, but they're bigger
tasks that we've lived with until now. All the major ones preventing doc
building or cross-referencing are now gone.

Reviewed at https://reviews.reviewboard.org/r/9238/