Make rbintegrations a packaging-only dependency.
In 3.0 beta 1, we attempted to make rbintegrations a dependency of
Review Board, which worked fine for the packaging process but created an
issue when building an environment to package rbintegrations. Since
rbintegrations was a hard dependency of Review Board, it created a
circular reference issue.

Now, we only include rbintegrations as a dependency if actively building
a package (through a `sdist` or `bdist` command), or if installing
(using `install`). Setting up a development environment no longer
requires it.

Testing Done:
Tested running `develop` and saw that `rbintegrations` wasn't a
dependency.

Tested the packaging commands and inspected the package metadata.
`rbintegrations` was a dependency.

Reviewed at https://reviews.reviewboard.org/r/9249/